requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
django-modeltranslation==0.19.16
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.10
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
